+++ orga/app/components/student/sup/header-actions.hbs
@@ -6,9 +6,9 @@
          href="{{this.urlToDownloadCsvTemplate}}" target="_blank" rel="noopener noreferrer" download>
         {{t "pages.students-sup.actions.download-template"}}
       </a>
-      <FileUpload @name="file-upload" @for="students-file-upload" @accept=".csv" @multiple={{false}} @onfileadd={{@onImportStudents}}>
-        <span class="button" role="button" tabindex="0">{{t "pages.students-sup.actions.import-file"}}</span>
-      </FileUpload>
+      <LinkTo @route="authenticated.sup-students.import" class="button button--link">
+        {{t 'pages.students-sup.actions.import-file'}}
+      </LinkTo>
     </div>
   {{/if}}
 </div>
