+++ orga/app/routes/authenticated/sup-students/import.js
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class ImportRoute extends Route {}
