+++ orga/app/templates/authenticated/sup-students/import.hbs
@@ -1,0 +1,12 @@
+{{page-title (t 'pages.students-sup.title')}}
+<div class="list-students-page">
+  <FileUpload @name="file-upload" @for="students-file-upload" @accept=".csv" @multiple={{false}} @onfileadd={{this.importStudents}}>
+    <span class="button" role="button" tabindex="0">{{t "pages.students-sup.actions.import-file"}}</span>
+  </FileUpload>
+</div>
+
+{{#if this.isLoading}}
+  <ModalDialog @translucentOverlay={{true}} @containerClass="modal__translucent">
+    <Ui::PixLoader />
+  </ModalDialog>
+{{/if}}
